Alteração dos tipos dos ids para Long

+++ account-management-api/src/main/java/com/pavarin/accountmanagementapi/entity/Account.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class Account {
 
-    int id;
-    double amount;
+    Long id;
+    Double amount;
     
 }
